Add deprecation template deprecated children in SwiftSyntaxBuilder
